[change] Renamed serializer to serializers in connection.api (consistency)
